feat: add user feature
