+++ src/components/dashboard/modals/index.ts
@@ -3,5 +3,3 @@ export * from './board/NewBoard'
 export * from './BurgerMenu'
 export * from './card/EditCard'
 export * from './card/NewCard'
-export * from './edit-profile/EditProfileModal'
-export * from './NeedHelp'
