+++ src/features/auth/auth.types.ts
@@ -1,4 +1,4 @@
-import type { User } from 'types'
+import type { User } from 'features/user/user.types'
 
 export type AuthResponse = {
   refreshToken: string
