+++ src/features/auth/hooks/useSigninUser.ts
@@ -1,5 +1,5 @@
-import type { SigninSchema } from 'lib/schemas'
 import type { UseFormReset } from 'react-hook-form'
+import type { SigninSchema } from '../auth.schema'
 
 import { useMutation } from '@tanstack/react-query'
 import { toast } from 'sonner'
