+++ src/features/user/components/index.ts
@@ -1,0 +1,2 @@
+export * from './EditProfileModal'
+export * from './NeedHelpModal'
