+++ src/features/user/config/user-api-endpoints.config.ts
@@ -1,0 +1,6 @@
+export const UserApiEndpoints = {
+  User: '/user',
+  UserMe: '/user/me',
+  UserTheme: '/user/theme',
+  UserHelp: '/user/help'
+}
