+++ src/features/user/config/user-cache-keys.config.ts
@@ -1,0 +1,6 @@
+export enum UserCacheKeys {
+  User = 'user',
+  UserHelp = 'userHelp',
+  EditUserProfile = 'editUserProfile',
+  ChangeUserTheme = 'changeUserTheme'
+}
