+++ src/features/user/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useChangeTheme'
 export * from './useEditProfile'
+export * from './useGetCurrentUser'
 export * from './useNeedHelp'
