+++ src/features/user/hooks/useEditProfile.ts
@@ -2,10 +2,9 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useModal } from 'react-modal-state'
 import { toast } from 'sonner'
 
-import { EditProfileModal } from 'components/dashboard/modals'
-
-import { CacheKeys } from 'config'
-import { userService } from 'services'
+import { EditProfileModal } from '../components'
+import { UserCacheKeys } from '../config'
+import { userService } from '../user.service'
 
 export const useEditProfile = () => {
   const queryClient = useQueryClient()
@@ -13,12 +12,12 @@ export const useEditProfile = () => {
   const { close } = useModal(EditProfileModal)
 
   return useMutation({
-    mutationKey: [CacheKeys.EditUserProfile],
+    mutationKey: [UserCacheKeys.EditUserProfile],
     mutationFn: userService.updateUserCredentials,
     onSuccess() {
       close()
       toast.success('Your profile has been successfully updated.')
-      queryClient.invalidateQueries({ queryKey: [CacheKeys.User] })
+      queryClient.invalidateQueries({ queryKey: [UserCacheKeys.User] })
     },
     onError(e) {
       toast.error(
