+++ src/index.tsx
@@ -4,6 +4,7 @@ import { StrictMode } from 'react'
 import { GoogleOAuthProvider } from '@react-oauth/google'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import { EditProfileModal, NeedHelpModal } from 'features/user/components'
 import ReactDOM from 'react-dom/client'
 import { ModalProvider, ModalRenderer } from 'react-modal-state'
 import { Provider } from 'react-redux'
@@ -15,8 +16,6 @@ import {
   BurgerMenu,
   EditBoardModal,
   EditCardModal,
-  EditProfileModal,
-  NeedHelpModal,
   NewBoardModal,
   NewCardModal
 } from 'components/dashboard/modals'
