+++ src/lib/schemas/index.ts
@@ -1,3 +1,2 @@
 export * from './board.schema'
 export * from './card.schema'
-export * from './user.schema'
