+++ src/types/index.ts
@@ -1,2 +1,1 @@
 export * from './board.types'
-export * from './user.types'
